requirements.txt
absl-py==1.4.0
addict==2.4.0
asciimatics==1.14.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chumpy==0.70
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.40.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h5py==3.9.0
hesiod==1.0.0
icecream==2.1.3
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.1
ipykernel==6.23.2
ipython==8.12.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kaolin @ file:///data_honghu/kaolin-0.14.0-cp38-cp38-linux_x86_64.whl#sha256=7b802b181d661625b68e838c814cba24a882f2343b1b71fe510bd12e673e4e5a
kiwisolver==1.4.4
lazy_loader==0.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
numpy==1.23.0
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.8.0.74
openmesh==1.2.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.6.0
plotly==5.15.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pyfiglet==0.8.post1
Pygments==2.15.1
pyhocon==0.3.60
PyMCubes==0.1.4
pyparsing==3.1.0
pyquaternion==0.9.9
pyrsistent==0.19.3
pysdf==0.1.8
python-dateutil==2.8.2
pytorch3d==0.7.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
siren-pytorch==0.1.6
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.7.10
torch @ https://download.pytorch.org/whl/cu113/torch-1.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=af11b85dfc75224198fe3693701c9488a7b6fcf9ef0df4c5c16949fd3ca8b89f
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.22.1
typeguard==2.13.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
usd-core==23.5
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
yacs==0.1.8
zipp==3.15.0
